imp: better break down of methods for basecoin
